backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
requests
email-validator
python-multipart
bcrypt==4.1.2
